Centralize the rootContainer and Worked on the FIlter Panel
Increased the width of the id yourDiv from 60% to 90%. This helped clear out the jumbled figures in the chart. 2. Centralize the rootContainer by applying the margin-left property. This should make it more visible and accessible. 3. Worked on the Filter Panel by: a. changing the background color and the text color. b. changing the Results length and c. applied border-radius accross all select and input tag.